build-linux-pkg.sh: Migrate from sudo to fakeroot
For every use of sudo in this script we can in fact use fakeroot. None of
these sudo wrappered commands really require root and best practice is
to fake it.

The default user permissions on Debian do not add the user to the
sudoers group. fakeroot is the right way to do this.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@nexus-software.ie>