Added support in roast properties to scan for bean color with lebrew roastsee c1 - with 2 buttons to the é fields with color - detect connection and disconnection, display a status and hide/show the button depending on detection - added device logic to define the lebrew rostsee c1 in network tab - added logic to save the settings in the alog file enhanced roast properties
Added the logic to define two additional comboboxes with ambiant information (pressure, humidity)
(these fields display the same popup of devices with extra devices than temperature). you can define
extra devices to add this additional fields.

Added an additional low priority thread to insure that the manchine is not going to screen saver or sleep mode
between ON and OFF operations. the monitoring is supported on windows, linux and MacOS (implements wakepy)